Multiple Linear Regression
Using the Boston house-prices dataset. The dataset consists of 13 features of 506 houses and the median home value in $1000's. First it is needed to fit a model on the 13 features to predict the value of the houses.

Afterwards, complete the following steps:

1. Build a linear regression model

Create a regression model using scikit-learn's LinearRegression and assign it to model.
Fit the model to the data.
2. Predict using the model

Predict the value of sample_house.